requirements.txt
appdirs==1.4.4
asgiref==3.6.0
astroid==2.4.2
attrs==22.2.0
autopep8==1.5.4
backports.zoneinfo==0.2.1
boto3==1.14.31
botocore==1.17.31
cachelib==0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
cs50==5.0.4
defusedxml==0.7.0rc1
distlib==0.3.1
Django==4.1.3
django-allauth==0.53.1
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-extensions==3.2.1
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.9.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docutils==0.15.2
exceptiongroup==1.1.1
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
gunicorn==20.0.4
idna==2.10
iniconfig==2.0.0
install==1.3.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==23.0
Pillow==8.0.1
pluggy==1.0.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
SQLAlchemy==1.3.17
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.2
virtualenv==20.2.1
Werkzeug==0.15.4
whitenoise==5.1.0
wrapt==1.12.1
